Move multiclass tests into test_engine.py
